Refactor ReactCrop to be used as a HOC (Higher Order Component)
ReactCrop was split into two parts, the HOC itself is now in "withCrop.js", while the default rendering component is in ReactCrop.  This should not be a breaking change, because the default export is going to be the ReactCrop component already wrapped in the HOC.
This HOC now allows developers more fine tuned control of the rendering process, so that it can be mixed with perhaps other HOC components that work with the same image/canvas.